adas_system/requirements.txt
# ADAS System Requirements

# Core dependencies
opencv-python==4.8.1
torch==2.1.0+cu121
torchvision==0.16.0+cu121
numpy==1.24.3
scipy==1.11.2
scikit-learn==1.3.1

# Object Detection & Vision
ultralytics==8.0.196  # YOLOv8
mmdet==3.1.0  # MMDetection
mmengine==0.8.4

# 3D Detection
mmcv==2.0.1
mmcv-full==1.7.4
open3d==0.17.0
pyquaternion==0.9.9

# Depth Estimation
monodepth2==0.1
timm==0.9.7  # Vision transformers

# Tracking
filterpy==1.4.2  # Kalman Filter
scikit-image==0.21.0  # DeepSORT dependencies
lap==0.4.0  # Hungarian algorithm

# Sensor Fusion & ROS2
rclpy==0.13.0
geometry_msgs==0.0.12
sensor_msgs==0.0.10
std_msgs==0.0.0

# Utilities
pyyaml==6.0.1
opencv-contrib-python==4.8.1
pillow==10.0.0
tqdm==4.66.1
attrs==23.1.0

# Optimization & Deployment
onnx==1.14.1
onnxruntime-gpu==1.16.0
tensorrt==8.6.1
pycuda==2021.1

# Data Processing
pandas==2.0.3
h5py==3.9.0
msgpack==1.0.7

# Development
jupyter==1.0.0
matplotlib==3.7.2
seaborn==0.12.2
tensorboard==2.13.0

# Visualization
rviz==2.16.5
